chore: move lib specific files to common/tools
